Refactored Service to make use of Connectors and Data modules
