+++ service/pom.xml
@@ -125,6 +125,18 @@
             <version>${kafka.version}</version>
             <scope>test</scope>
         </dependency>
+        <dependency>
+            <groupId>com.booking.validator</groupId>
+            <artifactId>connectors</artifactId>
+            <version>0.0.1</version>
+            <scope>compile</scope>
+        </dependency>
+        <dependency>
+            <groupId>com.booking.validator</groupId>
+            <artifactId>connectors</artifactId>
+            <version>0.0.1</version>
+            <scope>compile</scope>
+        </dependency>
 
     </dependencies>
