+++ service/src/test/java/FunctionalTest.java
@@ -3,8 +3,8 @@ import com.booking.validator.data.DataPointerFactory;
 import com.booking.validator.service.DataPointerFactories;
 import com.booking.validator.service.Validator;
 import com.booking.validator.service.TaskSupplier;
-import com.booking.validator.service.supplier.cli.CommandLineTaskSupplier;
-import com.booking.validator.service.supplier.kafka.KafkaTaskSupplier;
+import com.booking.validator.service.supplier.task.cli.CommandLineTaskSupplier;
+import com.booking.validator.service.supplier.task.kafka.KafkaTaskSupplier;
 import com.booking.validator.utils.Service;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
